requirements.txt
discord.py<=2
codingame[async]
jishaku
python-dotenv
sphobjinv[speedup]
